requirements-dev.txt
pytest
pypdf
pyyaml
requests
jsonschema
PyJWT
cryptography
